GV- hackathon/ai_module/requirements.txt
pymupdf
pdf2image
pillow
easyocr
torch
transformers
sentence-transformers
chromadb
langchain
pyttsx3
openai
requests
pytest
google-generativeai